requirements.txt
streamlit
pillow
tensorflow
opencv-python
numpy
pandas
googletrans==4.0.0-rc1